refactor: svg 경로 정리
