+++ src/components/layout/MyMarketFooterNav.tsx
@@ -3,11 +3,11 @@
 
 import React from 'react';
 import styled from 'styled-components';
-import HamburgerSvg from '/public/svgs/hamburger_icon.svg';
-import MyDongleSvg from '/public/svgs/mydongle_icon.svg';
-import DogHomeSvg from '/public/svgs/home_icon.svg';
-import BasketSvg from '/public/svgs/basket_icon_select.svg';
-import MySvg from '/public/svgs/user_icon.svg';
+import HamburgerSvg from '/public/svgs/navbar/hamburger_icon.svg';
+import MyDongleSvg from '/public/svgs/navbar/mydongle_icon.svg';
+import DogHomeSvg from '/public/svgs/navbar/home_icon.svg';
+import BasketSvg from '/public/svgs/navbar/basket_icon_select.svg';
+import MySvg from '/public/svgs/navbar/user_icon.svg';
 
 const FooterNavContainer = styled.nav`
   position: fixed;
